Reapply "Enable payload  availability check before voting"
This reverts commit a00795b2932ad4085d023d265a91397f0d405316.

+++ benchmark/benchmark/update_node.sh
@@ -18,9 +18,14 @@ BRANCH_NAME="$3"
 
 FUNC="update"
 
-eval $(ssh-agent)
-ssh-add /home/ubuntu/.ssh/"$DEPLOY_KEY_NAME"
-cd /home/ubuntu/"$REPO_NAME" 
+# Set up deploy key if provided and exists.
+if [ -n "$DEPLOY_KEY_NAME" ] && [ -f /home/ubuntu/.ssh/"$DEPLOY_KEY_NAME" ]; then
+    eval $(ssh-agent)
+    ssh-add /home/ubuntu/.ssh/"$DEPLOY_KEY_NAME"
+    HAVE_AGENT=1
+fi
+
+cd /home/ubuntu/"$REPO_NAME"
 git fetch -f
 git checkout -f "$BRANCH_NAME"
 git pull -f
@@ -32,9 +37,10 @@ cargo build --quiet --release --features benchmark
 cd /home/ubuntu
 rm -f node
 rm -f benchmark_client
-ln -s ./"$REPO_NAME"/target/release/node . 
+ln -s ./"$REPO_NAME"/target/release/node .
 ln -s ./"$REPO_NAME"/target/release/benchmark_client .
 
-kill "$SSH_AGENT_PID"
+# Cleanup ssh-agent (if started)
+[ -n "$HAVE_AGENT" ] && kill "$SSH_AGENT_PID"
 
 echo "$FUNC complete"
